// src/interfaces/IUserData.ts
export interface IUserData {
    email: string;
    username: string;
    id: string;
    is_mfaEnable: boolean;
    about: null | string;
    logo: null | string;
    name: null | string;
}

// about: null
// email: "mikalai.kozich@leverx.com"
// id: "1120ba66-1857-4951-9dfb-b01e69265669"
// is_mfa_enabled: false
// logo: null
// name: null
// username: "mikalai.kozich@leverx.com"